Move merlin.server.http.SerializedValueJsonParser to merlin.driver.json
